♻️ component에서 components폴더로 수정
