+++ src/components/Login/RememberMeToggle.jsx
@@ -1,7 +1,7 @@
-import { useState, createContext } from 'react';
-import useToggle from '../../hook/useToggle';
-import styled from 'styled-components';
-import COLOR from '../../style/color';
+import { useState, createContext } from "react";
+import useToggle from "../../hooks/useToggle";
+import styled from "styled-components";
+import COLOR from "../../style/color";
 
 export const ToggleContext = createContext();
 
@@ -10,24 +10,26 @@ const RememberMeToggle = ({ children }) => {
   const toggle = () => setOn(!on);
   return (
     <ToggleContext.Provider value={{ on, toggle }}>
-      <ToggleWrapper>
-        {children}
-      </ToggleWrapper>
-  </ToggleContext.Provider>);
+      <ToggleWrapper>{children}</ToggleWrapper>
+    </ToggleContext.Provider>
+  );
 };
 
 const ToggleState = () => {
   const { on } = useToggle();
-  return <ToggleStateBox on={on|false}>&nbsp;{on ? 'ON' : 'OFF'}</ToggleStateBox>
-}
+  return (
+    <ToggleStateBox on={on | false}>&nbsp;{on ? "ON" : "OFF"}</ToggleStateBox>
+  );
+};
 
 const ToggleButton = () => {
   const { on, toggle } = useToggle();
   return (
     <ToggleButtonContainer>
-      <ToggleInput type='checkbox' checked={on|false} onChange={toggle} />
+      <ToggleInput type="checkbox" checked={on | false} onChange={toggle} />
       <ToggleSlider></ToggleSlider>
-    </ToggleButtonContainer>);
+    </ToggleButtonContainer>
+  );
 };
 
 RememberMeToggle.ToggleState = ToggleState;
@@ -39,30 +41,30 @@ const ToggleWrapper = styled.div`
   display: flex;
   justify-content: left;
   align-items: center;
-`
+`;
 
 const ToggleStateBox = styled.span`
-  ${({ on }) => on? `color: ${COLOR.MAIN};` : `color: lightgray;` }
+  ${({ on }) => (on ? `color: ${COLOR.MAIN};` : `color: lightgray;`)}
   font-weight: bold;
-`
+`;
 
 const ToggleButtonContainer = styled.label`
   display: inline-block;
   position: relative;
   width: 2.6rem;
   height: 1.5rem;
-`
+`;
 const ToggleInput = styled.input`
   width: 0;
   height: 0;
   opacity: 0;
-  &:checked + span{
+  &:checked + span {
     background-color: ${COLOR.MAIN};
   }
-  &:checked + span::before{
+  &:checked + span::before {
     transform: translateX(1.1rem);
   }
-`
+`;
 
 const ToggleSlider = styled.span`
   position: absolute;
@@ -72,9 +74,9 @@ const ToggleSlider = styled.span`
   bottom: 0;
   border-radius: 1.7rem;
   background-color: lightgray;
-  transition: .4s;
+  transition: 0.4s;
   cursor: pointer;
-  &::before{
+  &::before {
     position: absolute;
     left: 0.2rem;
     bottom: 0.2rem;
@@ -83,7 +85,6 @@ const ToggleSlider = styled.span`
     border-radius: 50%;
     background-color: white;
     content: "";
-    transition: .4s;
+    transition: 0.4s;
   }
-
-`
+`;
