+++ src/hooks/useToggle.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { ToggleContext } from "../components/Login/RememberMeToggle";
+
+const useToggle = () => {
+  const context = React.useContext(ToggleContext);
+  if (context === undefined)
+    throw new Error("useToggle must be used within a <Toggle />");
+  return context;
+};
+
+export default useToggle;
